agrolink-backend/requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai
scikit-learn
pandas
numpy
joblib
torch
transformers
sentencepiece
firebase-admin
python-multipart
Pillow
